requirements.txt
networkx==3.2.1
numpy==1.26.4
python-igraph==0.11.9
pandas==2.2.2
scikit-learn==1.3.2
